refactor: self review comments
